backend/requirements.txt
flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.0
firebase-admin==6.2.0
flask-cors==4.0.0
python-dateutil==2.8.2
requests==2.31.0
google-cloud-aiplatform==1.36.4
pypdf2==3.0.1
python-jose==3.3.0
gunicorn==21.2.0
